+++ frontend/schemaCMS/src/routes/project/dataSourceList/__tests__/dataSourceList.component.spec.js
@@ -3,8 +3,6 @@ import { shallow } from 'enzyme';
 
 import { DataSourceList } from '../dataSourceList.component';
 import { defaultProps, propsWithDataSources } from '../dataSourceList.stories';
-import { PlusButton } from '../../../../shared/components/navigation';
-import { ListItemTitle } from '../../../../shared/components/listComponents';
 
 describe('DataSourceList: Component', () => {
   const component = props => <DataSourceList {...defaultProps} {...props} />;
@@ -19,54 +17,6 @@ describe('DataSourceList: Component', () => {
   it('should render correctly', async () => {
     const wrapper = await render(propsWithDataSources);
     expect(wrapper).toMatchSnapshot();
-  });
-
-  it('should go to source', async () => {
-    jest.spyOn(propsWithDataSources.history, 'push');
-    const wrapper = await render(propsWithDataSources);
-
-    wrapper
-      .find(ListItemTitle)
-      .at(0)
-      .simulate('click');
-
-    expect(propsWithDataSources.history.push).toHaveBeenCalledWith('/datasource/17/source');
-  });
-
-  it('should go to preview', async () => {
-    jest.spyOn(propsWithDataSources.history, 'push');
-    const wrapper = await render(propsWithDataSources);
-
-    wrapper
-      .find(ListItemTitle)
-      .at(1)
-      .simulate('click');
-
-    expect(propsWithDataSources.history.push).toHaveBeenCalledWith('/datasource/17/preview');
-  });
-
-  it('should go to data result', async () => {
-    jest.spyOn(propsWithDataSources.history, 'push');
-    const wrapper = await render(propsWithDataSources);
-
-    wrapper
-      .find(ListItemTitle)
-      .at(2)
-      .simulate('click');
-
-    expect(propsWithDataSources.history.push).toHaveBeenCalledWith('/datasource/17/result');
-  });
-
-  it('should go to create data source page', () => {
-    jest.spyOn(propsWithDataSources.history, 'push');
-    const wrapper = render(propsWithDataSources);
-
-    wrapper
-      .find(PlusButton)
-      .at(0)
-      .simulate('click');
-
-    expect(propsWithDataSources.history.push).toHaveBeenCalledWith('/project/1/datasource/add');
   });
 
   it('should call fetchDataSources on componentDidMount', async () => {