+++ frontend/schemaCMS/src/routes/project/tagCategories/tagCategories.component.js
@@ -7,7 +7,7 @@ import { camelize } from 'humps';
 import messages from './tagCategories.messages';
 import { LoadingWrapper } from '../../../shared/components/loadingWrapper';
 import { ContextHeader } from '../../../shared/components/contextHeader';
-import { BackArrowButton, NavigationContainer, PlusButton } from '../../../shared/components/navigation';
+import { NavigationContainer, PlusLink, LARGE_BUTTON_SIZE, BackLink } from '../../../shared/components/navigation';
 import { filterMenuOptions, getMatchParam } from '../../../shared/utils/helpers';
 import reportError from '../../../shared/utils/reportError';
 import { TAG_CATEGORIES_PAGE } from '../../../modules/project/project.constants';
@@ -79,24 +79,20 @@ export class TagCategories extends PureComponent {
     }
   }
 
-  handleCreateTag = () => {
+  getCreateUrl = () => {
     const projectId = getMatchParam(this.props, 'projectId');
-    this.props.history.push(`/project/${projectId}/${TAG_CATEGORIES_PAGE}/create`);
+    return `/project/${projectId}/${TAG_CATEGORIES_PAGE}/create`;
   };
 
-  handleShowTemplates = () => this.props.history.push(`/project/${getMatchParam(this.props, 'projectId')}/templates`);
-
   renderTagCategory = ({ created, createdBy, name, id, tags }, index) => {
-    const { history } = this.props;
     const whenCreated = extendedDayjs(created, BASE_DATE_FORMAT).fromNow();
     const list = [whenCreated, createdBy];
     const header = <CardHeader list={list} />;
     const footer = <FormattedMessage {...messages.tagsCounter} values={{ count: tags.length }} />;
-    const handlePageClick = () => history.push(`/tag-category/${id}`);
 
     return (
       <ListItem id="tagContainer" headerComponent={header} footerComponent={footer} key={index}>
-        <ListItemTitle id={`tag-category-${camelize(name)}`} onClick={handlePageClick}>
+        <ListItemTitle id={`tag-category-${camelize(name)}`} to={`/tag-category/${id}`}>
           {name}
         </ListItemTitle>
       </ListItem>
@@ -110,8 +106,8 @@ export class TagCategories extends PureComponent {
       <Fragment>
         <ListContainer>{tagCategories.map(this.renderTagCategory)}</ListContainer>
         <NavigationContainer fixed>
-          <BackArrowButton id="backBtn" hideOnDesktop onClick={this.handleShowTemplates} />
-          <PlusButton hideOnDesktop onClick={this.handleCreateTag} />
+          <BackLink id="backBtn" to={`/project/${getMatchParam(this.props, 'projectId')}/templates`} />
+          <PlusLink hideOnDesktop to={this.getCreateUrl()} size={LARGE_BUTTON_SIZE} />
         </NavigationContainer>
       </Fragment>
     );
@@ -134,7 +130,7 @@ export class TagCategories extends PureComponent {
           options={filterMenuOptions(menuOptions, userRole)}
         />
         <ContextHeader title={headerTitle} subtitle={headerSubtitle}>
-          <PlusButton id="addTagCategoryBtn" onClick={this.handleCreateTag} />
+          <PlusLink id="addTagCategoryBtn" to={this.getCreateUrl()} size={LARGE_BUTTON_SIZE} />
         </ContextHeader>
         <LoadingWrapper loading={loading} error={error}>
           {this.renderContent()}