+++ frontend/schemaCMS/src/routes/project/userDetails/__tests__/userDetails.component.spec.js
@@ -3,7 +3,7 @@ import { shallow } from 'enzyme';
 
 import { UserDetails } from '../userDetails.component';
 import { defaultProps } from '../userDetails.stories';
-import { BackButton, NextButton } from '../../../../shared/components/navigation';
+import { NextButton } from '../../../../shared/components/navigation';
 
 describe('UserDetails: Component', () => {
   const component = props => <UserDetails {...defaultProps} {...props} />;
@@ -29,18 +29,6 @@ describe('UserDetails: Component', () => {
     render();
 
     expect(defaultProps.fetchUser).toBeCalledWith({ userId: '1' });
-  });
-
-  it('should go back', () => {
-    jest.spyOn(defaultProps.history, 'push');
-
-    const wrapper = render(defaultProps);
-    wrapper
-      .find(BackButton)
-      .at(0)
-      .simulate('click');
-
-    expect(defaultProps.history.push).toBeCalledWith('/project/1/user');
   });
 
   it('should remove user', () => {