+++ frontend/schemaCMS/src/shared/components/userCreate/userCreateCMS/__tests__/userCreateCMS.spec.js
@@ -4,8 +4,6 @@ import { Formik } from 'formik';
 
 import { UserCreateCMS } from '../userCreateCMS.component';
 import { userCreateCMSProps } from '../userCreateCMS.stories';
-import { UserCreate } from '../../userCreateComponent/userCreate.component';
-import browserHistory from '../../../../utils/history';
 
 describe('UserCreateCMS: Component', () => {
   const component = props => <UserCreateCMS {...userCreateCMSProps} {...props} />;
@@ -28,19 +26,5 @@ describe('UserCreateCMS: Component', () => {
     );
 
     expect(userCreateCMSProps.createUserCMS).toHaveBeenCalledTimes(1);
-  });
-
-  it('should redirect on cancel', async () => {
-    jest.spyOn(browserHistory, 'push');
-
-    const wrapper = render();
-
-    wrapper
-      .dive()
-      .dive()
-      .find(UserCreate)
-      .prop('onCancelClick')({ preventDefault: Function.prototype });
-
-    expect(browserHistory.push).toHaveBeenCalledWith('/user');
   });
 });